+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * #2147: Fix piston platforms not killing objects when retracting.
 * Fix the infinite value not working in any item properties options.
 * Ensure signage reconfiguration applies to exports even if window is not closed.
+* Add support for barrier holes on "Start Reversed" offset glass/grating.
 
 # Version 4.46.1
 * Fix FGD export producing invalid syntax.